requirements/base.txt
# Base Requirements - Shared across all models
# All versions pinned for reproducibility

# Core Orchestration
apache-airflow==2.8.1

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Configuration Management
pyyaml==6.0.1

# Machine Learning Utilities
scikit-learn==1.3.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
colorlog==6.7.0
